requirements.txt
fastapi
uvicorn
motor
pydantic[email]
bcrypt
python-dotenv
python-multipart